Fix module path: core/rules/business → core/business/rules
Corrects the module and package structure to core/business/rules/{api,impl}
with package cz.adamec.timotej.snag.core.business.rules.
