+++ build-logic/src/main/kotlin/cz/adamec/timotej/snag/buildsrc/configuration/AutoWiring.kt
@@ -98,7 +98,7 @@ internal fun Project.resolveHexagonalDependencies(): List<AutoWiredDependency> {
 
     // shared validation rules for all feat modules
     if (path.startsWith(":feat:")) {
-        result += AutoWiredDependency(":core:rules:business:api", DependencyScope.IMPLEMENTATION)
+        result += AutoWiredDependency(":core:business:rules:api", DependencyScope.IMPLEMENTATION)
     }
 
     return result
