+++ core/business/rules/api/src/commonMain/kotlin/cz/adamec/timotej/snag/core/business/rules/api/PhoneNumberRule.kt
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) 2026 Timotej Adamec
+ * SPDX-License-Identifier: MIT
+ *
+ * This file is part of the thesis:
+ * "Multiplatform snagging system with code sharing maximisation"
+ *
+ * Czech Technical University in Prague
+ * Faculty of Information Technology
+ * Department of Software Engineering
+ */
+
+package cz.adamec.timotej.snag.core.business.rules.api
+
+interface PhoneNumberRule {
+    operator fun invoke(phoneNumber: String): Boolean
+}
