+++ core/business/rules/impl/src/commonMain/kotlin/cz/adamec/timotej/snag/core/business/rules/impl/internal/EmailFormatRuleImpl.kt
@@ -1,0 +1,23 @@
+/*
+ * Copyright (c) 2026 Timotej Adamec
+ * SPDX-License-Identifier: MIT
+ *
+ * This file is part of the thesis:
+ * "Multiplatform snagging system with code sharing maximisation"
+ *
+ * Czech Technical University in Prague
+ * Faculty of Information Technology
+ * Department of Software Engineering
+ */
+
+package cz.adamec.timotej.snag.core.business.rules.impl.internal
+
+import cz.adamec.timotej.snag.core.business.rules.api.EmailFormatRule
+
+internal class EmailFormatRuleImpl : EmailFormatRule {
+    override operator fun invoke(email: String): Boolean = email.matches(EMAIL_REGEX)
+
+    private companion object {
+        val EMAIL_REGEX = Regex("^[A-Za-z0-9+_.-]+@[A-Za-z0-9]+([.-][A-Za-z0-9]+)*\\.[A-Za-z]{2,}$")
+    }
+}
