--- core/rules/business/impl/build.gradle.kts
@@ -1,23 +1,0 @@
-/*
- * Copyright (c) 2026 Timotej Adamec
- * SPDX-License-Identifier: MIT
- *
- * This file is part of the thesis:
- * "Multiplatform snagging system with code sharing maximisation"
- *
- * Czech Technical University in Prague
- * Faculty of Information Technology
- * Department of Software Engineering
- */
-
-plugins {
-    alias(libs.plugins.snagMultiplatformModule)
-}
-
-kotlin {
-    sourceSets {
-        commonMain.dependencies {
-            implementation(projects.core.rules.business.api)
-        }
-    }
-}
