--- core/rules/business/impl/src/commonMain/kotlin/cz/adamec/timotej/snag/core/rules/business/impl/internal/PhoneNumberRuleImpl.kt
@@ -1,23 +1,0 @@
-/*
- * Copyright (c) 2026 Timotej Adamec
- * SPDX-License-Identifier: MIT
- *
- * This file is part of the thesis:
- * "Multiplatform snagging system with code sharing maximisation"
- *
- * Czech Technical University in Prague
- * Faculty of Information Technology
- * Department of Software Engineering
- */
-
-package cz.adamec.timotej.snag.core.rules.business.impl.internal
-
-import cz.adamec.timotej.snag.core.rules.business.api.PhoneNumberRule
-
-internal class PhoneNumberRuleImpl : PhoneNumberRule {
-    override operator fun invoke(phoneNumber: String): Boolean =
-        phoneNumber == phoneNumber.trim() &&
-            phoneNumber.isNotBlank() &&
-            phoneNumber.none { it.isLetter() } &&
-            phoneNumber.any { it.isDigit() }
-}
