+++ feat/clients/fe/driving/impl/src/commonMain/kotlin/cz/adamec/timotej/snag/clients/fe/driving/impl/internal/clientDetailsEdit/vm/ClientDetailsEditViewModel.kt
@@ -17,9 +17,9 @@ import androidx.lifecycle.viewModelScope
 import cz.adamec.timotej.snag.clients.fe.app.api.GetClientUseCase
 import cz.adamec.timotej.snag.clients.fe.app.api.SaveClientUseCase
 import cz.adamec.timotej.snag.clients.fe.app.api.model.SaveClientRequest
+import cz.adamec.timotej.snag.core.business.rules.api.EmailFormatRule
+import cz.adamec.timotej.snag.core.business.rules.api.PhoneNumberRule
 import cz.adamec.timotej.snag.core.network.fe.OfflineFirstDataResult
-import cz.adamec.timotej.snag.core.rules.business.api.EmailFormatRule
-import cz.adamec.timotej.snag.core.rules.business.api.PhoneNumberRule
 import cz.adamec.timotej.snag.lib.design.fe.error.UiError
 import cz.adamec.timotej.snag.lib.design.fe.error.UiError.Unknown
 import kotlinx.coroutines.cancel
