+++ settings.gradle.kts
@@ -75,8 +75,8 @@ include(":feat:shared:database:fe:impl")
 include(":feat:shared:database:fe:test")
 include(":feat:shared:database:be:impl")
 include(":feat:shared:database:be:test")
-include(":core:rules:business:api")
-include(":core:rules:business:impl")
+include(":core:business:rules:api")
+include(":core:business:rules:impl")
 include(":lib:design:fe")
 include(":lib:routing:fe")
 include(":lib:routing:be")
